# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "这是一个用 Go 编写的博客后端 API 示例，支持用户、博客、评论、点赞等功能。",
        "title": "博客后端 API 文档",
        "contact": {
            "name": "API Support",
            "email": "youremail@example.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/admin/blog": {
            "post": {
                "description": "创建新博客文章",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "博客"
                ],
                "summary": "创建新博客文章",
                "parameters": [
                    {
                        "description": "博客内容",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateBlogRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.BlogResponse"
                        }
                    },
                    "400": {
                        "description": "无效的请求数据",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "未认证用户",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "创建文章失败",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/admin/blog/{id}": {
            "put": {
                "description": "更新博客文章",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "博客"
                ],
                "summary": "更新博客文章",
                "parameters": [
                    {
                        "type": "string",
                        "description": "博客ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新内容",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateBlogRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.BlogResponse"
                        }
                    },
                    "400": {
                        "description": "无效的请求数据",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "未认证用户",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "文章未找到",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除博客文章",
                "tags": [
                    "博客"
                ],
                "summary": "删除博客文章",
                "parameters": [
                    {
                        "type": "string",
                        "description": "博客ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "删除成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "未认证用户",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "文章未找到",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/admin/tag": {
            "post": {
                "description": "新增一个标签（如有专门标签集合时用）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "博客"
                ],
                "summary": "新增标签",
                "parameters": [
                    {
                        "description": "标签内容",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.TagRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "无效的标签",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "添加标签失败",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除标签（会从所有博客中移除该标签）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "博客"
                ],
                "summary": "删除标签",
                "parameters": [
                    {
                        "description": "标签内容",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.TagRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "删除成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "无效的标签",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "删除标签失败",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/blog/{id}": {
            "get": {
                "description": "根据 ID 获取博客详情",
                "tags": [
                    "博客"
                ],
                "summary": "获取单篇博客文章",
                "parameters": [
                    {
                        "type": "string",
                        "description": "博客ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.BlogResponse"
                        }
                    },
                    "404": {
                        "description": "文章未找到",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/blogs": {
            "get": {
                "description": "分页获取博客文章",
                "tags": [
                    "博客"
                ],
                "summary": "分页获取博客文章",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.BlogListResponse"
                        }
                    },
                    "500": {
                        "description": "获取文章失败",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/blogs/by-tags": {
            "get": {
                "description": "根据标签数组分页获取所有包含这些标签的博客文章",
                "tags": [
                    "博客"
                ],
                "summary": "分页获取包含所有指定标签的文章",
                "parameters": [
                    {
                        "type": "string",
                        "description": "标签（可重复）",
                        "name": "tag",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.BlogListResponse"
                        }
                    },
                    "500": {
                        "description": "获取文章失败",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/blogs/search": {
            "get": {
                "description": "支持关键字和标签数组的分页搜索",
                "tags": [
                    "博客"
                ],
                "summary": "博客模糊搜索与标签筛选",
                "parameters": [
                    {
                        "type": "string",
                        "description": "关键字",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "标签（可重复）",
                        "name": "tag",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.BlogListResponse"
                        }
                    },
                    "500": {
                        "description": "搜索失败",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/tags": {
            "get": {
                "description": "获取所有博客标签",
                "tags": [
                    "博客"
                ],
                "summary": "获取所有标签",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "获取标签失败",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handlers.BlogListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Blog"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/handlers.Pagination"
                }
            }
        },
        "handlers.BlogResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Blog"
                }
            }
        },
...
        "handlers.Pagination": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "handlers.TagRequest": {
            "type": "object",
            "properties": {
                "tag": {
                    "type": "string"
                }
            }
        },
...
        "models.Blog": {
            "type": "object",
            "properties": {
                "author": {
                    "description": "博客文章的作者",
                    "type": "string"
                },
                "content": {
                    "description": "博客文章的内容",
                    "type": "string"
                },
                "created_at": {
                    "description": "博客文章的创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "博客文章的唯一标识符",
                    "type": "string"
                },
                "show": {
                    "description": "是否在前端展示",
                    "type": "boolean"
                },
                "tags": {
                    "description": "内标签数组",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "description": "博客文章的标题",
                    "type": "string"
                },
                "updated_at": {
                    "description": "博客文章的更新时间",
                    "type": "string"
                },
                "views": {
                    "description": "浏览次数",
                    "type": "integer"
                }
            }
        },
...
        "models.CreateBlogRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "show": {
                    "type": "boolean"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "models.UpdateBlogRequest": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "show": {
                    "type": "boolean"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "views": {
                    "type": "integer"
                }
            }
        },
...
    }
}