# models/blog.go
package models

import (
	"time"

	"go.mongodb.org/mongo-driver/bson/primitive"
)

// Blog 表示一篇博客文章
type Blog struct {
	ID        primitive.ObjectID `bson:"_id,omitempty" json:"id,omitempty"`    // 博客文章的唯一标识符
	Title     string             `bson:"title" json:"title"`                   // 博客文章的标题
	Content   string             `bson:"content" json:"content"`               // 博客文章的内容
	Author    string             `bson:"author" json:"author"`                 // 博客文章的作者
	Tags      []string           `bson:"tags,omitempty" json:"tags,omitempty"` // 内标签数组
	Views     int64              `bson:"views" json:"views"`                   // 浏览次数
	Show      bool               `bson:"show" json:"show"`                     // 是否在前端展示
	CreatedAt time.Time          `bson:"created_at" json:"created_at"`         // 博客文章的创建时间
	UpdatedAt time.Time          `bson:"updated_at" json:"updated_at"`         // 博客文章的更新时间
}

// CreateBlogRequest 创建博客请求
type CreateBlogRequest struct {
	Title   string   `json:"title"`
	Content string   `json:"content"`
	Tags    []string `json:"tags,omitempty"`
	Show    *bool    `json:"show,omitempty"`
}

// UpdateBlogRequest 更新博客请求
type UpdateBlogRequest struct {
	Title   *string  `json:"title,omitempty"`
	Content *string  `json:"content,omitempty"`
	Author  *string  `json:"author,omitempty"`
	Tags    []string `json:"tags,omitempty"`
	Show    *bool    `json:"show,omitempty"`
	Views   *int64   `json:"views,omitempty"`
}